pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask",
    "gunicorn",
    "python-dotenv",
    "psycopg2-binary",
    "validators",
    "requests",
    "ruff",
    "beautifulsoup4"
]

[project.optional-dependencies]
test = ["pytest"]